+++ castor.php
@@ -2,8 +2,20 @@
 
 defined('CASTOR_USE_CHDIR') || define('CASTOR_USE_CHDIR', true);
 
+use Castor\Attribute\AsContext;
+use Castor\Context;
+
 use function Castor\import;
-use function Castor\mount;
 
 import(__DIR__ . '/.castor');
-mount(__DIR__ . '/demo', 'demo');
+
+#[AsContext(default: true)]
+function default_context(): Context
+{
+    return new Context([
+        'project_name' => 'jolimediabundle-demo',
+        'root_domain' => 'jolimediabundle-demo.test',
+        // used by "castor demo:docker:push"
+        'registry' => 'ghcr.io/jolicode/media-bundle',
+    ]);
+}