+++ demo/application/rector.php
@@ -3,14 +3,11 @@
 use Rector\Caching\ValueObject\Storage\FileCacheStorage;
 use Rector\Config\RectorConfig;
 use Rector\Doctrine\Set\DoctrineSetList;
-use Rector\Symfony\Configs\Rector\Closure\FromServicePublicToDefaultsPublicRector;
-use Rector\Symfony\Configs\Rector\Closure\ServiceTagsToDefaultsAutoconfigureRector;
 use Rector\Symfony\Set\SymfonySetList;
-use Rector\Symfony\Set\TwigSetList;
 use Rector\TypeDeclaration\Rector\StmtsAwareInterface\SafeDeclareStrictTypesRector;
 
 return RectorConfig::configure()
-    ->withCache('../tools/rector/var/cache', FileCacheStorage::class)
+    ->withCache(__DIR__ . '/var/cache/rector', FileCacheStorage::class)
     ->withPaths([
         __DIR__ . '/config',
         __DIR__ . '/migrations',
@@ -28,19 +25,14 @@ return RectorConfig::configure()
         naming: false,
         typeDeclarations: true,
     )
+    // the Symfony and Twig upgrade sets are picked from the versions in composer.json
+    ->withComposerBased(twig: true, doctrine: true, symfony: true)
     ->withSets([
         DoctrineSetList::DOCTRINE_CODE_QUALITY,
-        SymfonySetList::CONFIGS,
-        SymfonySetList::SYMFONY_70,
-        SymfonySetList::SYMFONY_71,
-        SymfonySetList::SYMFONY_72,
         SymfonySetList::SYMFONY_CODE_QUALITY,
         SymfonySetList::SYMFONY_CONSTRUCTOR_INJECTION,
-        TwigSetList::TWIG_UNDERSCORE_TO_NAMESPACE,
     ])
     ->withSkip([
-        ServiceTagsToDefaultsAutoconfigureRector::class,
-        FromServicePublicToDefaultsPublicRector::class,
         SafeDeclareStrictTypesRector::class,
         // Flex-managed file, keep the inline fully qualified class names
         __DIR__ . '/config/bundles.php',